include: Fix _memccpy, _memicmp and _memicmp_l prototypes.

+++ include/msvcrt/corecrt_wstring.h
@@ -19,9 +19,9 @@ _ACRTIMP int     __cdecl memcmp(const void*,const void*,size_t);
 _ACRTIMP void*   __cdecl memcpy(void*,const void*,size_t);
 _ACRTIMP errno_t __cdecl memcpy_s(void*,size_t,const void*,size_t);
 _ACRTIMP void*   __cdecl memset(void*,int,size_t);
-_ACRTIMP void*   __cdecl _memccpy(void*,const void*,int,unsigned int);
-_ACRTIMP int     __cdecl _memicmp(const void*,const void*,unsigned int);
-_ACRTIMP int     __cdecl _memicmp_l(const void*,const void*,unsigned int,_locale_t);
+_ACRTIMP void*   __cdecl _memccpy(void*,const void*,int,size_t);
+_ACRTIMP int     __cdecl _memicmp(const void*,const void*,size_t);
+_ACRTIMP int     __cdecl _memicmp_l(const void*,const void*,size_t,_locale_t);
 
 static inline int memicmp(const void* s1, const void* s2, size_t len) { return _memicmp(s1, s2, len); }
 static inline void* memccpy(void *s1, const void *s2, int c, size_t n) { return _memccpy(s1, s2, c, n); }
